// src/pages/WebSQL/type.d.ts
interface DOMString extends String { }
interface ObjectArray extends Array<any> { }

// [Supplemental, NoInterfaceObject]
export declare interface WindowDatabase {
  openDatabase (
    name: DOMString,
    version: DOMString,
    displayName: DOMString,
    estimatedSize: number,
    creationCallback?: DatabaseCallback,
  ): Database;
}
// ...
interface DatabaseCallback {
    /*handleEvent*/ (database: Database): void;
}
// ...
interface Database {
    /*readonly/const*/ version: DOMString;

  transaction (
    callback: SQLTransactionCallback,
    errorCallback?: SQLTransactionErrorCallback,
    successCallback?: SQLVoidCallback,
  ): void;

  readTransaction (
    callback: SQLTransactionCallback,
    errorCallback?: SQLTransactionErrorCallback,
    successCallback?: SQLVoidCallback,
  ): void;

  /** The changeVersion() method allows scripts to atomically verify the version number and change
   * it at the same time as doing a schema update. When the method is invoked, it must immediately
   * return, and then asynchronously run the transaction steps with the transaction callback being
   * the third argument, the error callback being the fourth argument, the success callback being
   * the fifth argument
   */
  changeVersion (
    oldVersion: DOMString,
    newVersion: DOMString,
    callback?: SQLTransactionCallback,
    errorCallback?: SQLTransactionErrorCallback,
    successCallback?: SQLVoidCallback,
  ): void;
}

// [Callback = FunctionOnly, NoInterfaceObject]
interface SQLVoidCallback {
    /*handleEvent*/ (): void;
}

// [Callback = FunctionOnly, NoInterfaceObject]
interface SQLTransactionCallback {
    /*handleEvent*/ (transaction: SQLTransaction): void;
}

// [Callback = FunctionOnly, NoInterfaceObject]
interface SQLTransactionErrorCallback {
    /*handleEvent*/ (error: SQLError): void;
}
// ...
export interface SQLResultSet {
  insertId: number;
  rowsAffected: number;
  rows: SQLResultSetRowList;
}
// ...
export interface SQLResultSetRowList {
  length: number;
  item (index: number): any;
}